Add serialization and QueryOption option

+++ warehouse/query-core/src/main/java/datawave/query/config/ShardQueryConfiguration.java
@@ -1571,7 +1571,7 @@ public class ShardQueryConfiguration extends GenericQueryConfiguration implement
     }
 
     public int getMaxIvaratorSources() {
-        return ivaratorConfig.getMaxIvaratorSources();
+        return (int)ivaratorConfig.getMaxIvaratorSources();
     }
 
     public void setMaxIvaratorSources(int maxIvaratorSources) {
